create table parcial

+++ internal/domain/Categoria.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Categoria struct {
-	IDCategoria int64  `db:"id_categoria" json:"idCategoria"`
+	IDCategoria int    `db:"id_categoria" json:"idCategoria"`
 	Nombre      string `db:"nombre"        json:"nombre"`
 	Descripcion string `db:"descripcion"   json:"descripcion"`
 }
